* configure.in (AC_OUTPUT): Build subversion/libsvn_fs/tests/Makefile. * subversion/libsvn_fs/Makefile.am (SUBDIRS): Mention `tests'.
* subversion/include/svn_fs.h (svn_fs_t): The underlying structure for
this typedef is now named `struct svn_fs_t', not `struct svn_fs', for
consistency.
* subversion/libsvn_fs/fs.h (struct svn_fs_t): Renamed from `struct svn_fs'.

* subversion/include/svn_fs.h: Doc fixes.

* subversion/libsvn_fs/Makefile.am (libsvn_fs_la_SOURCES): Remove
delta.c; it's incomplete.
* subversion/libsvn_fs/delta.c: Reintroduce syntax errors; this code
isn't runnable anyway, and the syntax errors ensure that I won't
forget the parts that are incomplete.  Since this file isn't mentioned
in the Makefile anyway, this shouldn't cause anyone problems.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@837434 13f79535-47bb-0310-9956-ffa450edef68