(make_dir_baton): revert back to apr_pcalloc().
[Note to Greg H: totally a matter of personal taste how you do it in
code you're working on, but for code I "own" -- in the loose sense of
ownership, of course -- I'd like to stick to the pcalloc way.  In this
case, your change was not harmless, because d->prop_changed would now
have a garbage value, instead of the zero I was counting on from
pcalloc.  I know you don't favor this way of doing initialization, but
I personally find it less distracting, and would like to use it where
I'm working.  Thanks.  -K]


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@837462 13f79535-47bb-0310-9956-ffa450edef68