* include/svn_fs.h: Miscellaneous changes throughout, as we figure out what form the filesystem should really have.  Biggest changes:
- Rework pool system.  Allocate all immutable nodes in the
filesystem's pool, and all mutable nodes in the transaction's pool,
but make it easy for people to register cleanup functions with their
own pools that close nodes.  Where we do let the user specify pools,
don't require them to be subpools of the filesystem; just make the
user responsible for freeing them before the filesystem is closed.

- Make the filesystem open / create functions Berkeley-DB specific.
This doesn't lose any real generality, and it means that we can add
functions to open / create other kinds of filesystems without losing
symmetry.

- Add new generic node type, svn_fs_node_t, which acts as a common
superclass for both files and directories.  This allows us to declare
functions that operate on any kind of node, instead of having to write
out separate functions for files and directories.  For example, we've
combined `open_file' and `open_dir' into `open_node'.

- When adding and replacing files and directories, use immutable node
objects to specify the base version, not filenames and version numbers.

* libsvn_fs: Much work throughout.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@837499 13f79535-47bb-0310-9956-ffa450edef68