Clean up some commentary.
The svn client main procedure has now been rewritten to use a command
dispatch/descriptor table.  That table plus the `help' command is the
only command specific information that should be in the main.c file.
The option processing is creaky, still there and on intimate terms with
the various commands.  It will go away RSN and be replaced with some
sort of table-driven version.  Each command will have its own option
table and that table will be referenced from the command descriptor table.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@838017 13f79535-47bb-0310-9956-ffa450edef68