Patch by Mo DeJong  <mdejong@redhat.com>, reviewed and applied as is.
Removed a block in main that was adding a "." argument if the command
type is one of the commands that takes an implicit dot. This seems to
be a bad idea since the main() function needs to know about every
possible command and it is easy to miss one. For example, the diff
command is not included in this list.  It really should be, `svn diff`
should really work the same as `cvs diff`. It seems like a much better
solution would be to do this implicit dot check inside the functions
that actually need to do it.

   * subversion/client/cl.h (push_implicit_dot_target):
   Add header for new method defined in main.c.
   * subversion/client/commit-cmd.c (svn_cl__commit):
   Call push_implicit_dot_target.
   * subversion/client/diff-cmd.c (svn_cl__diff):
   Call push_implicit_dot_target.
   * subversion/client/main.c: #include <assert.h>,
   define push_implicit_dot_target. Remove error
   prone block of code to push implicit dot target
   by comparing to known commands.
   * subversion/client/propget-cmd.c (svn_cl__propget):
   * subversion/client/proplist-cmd.c (svn_cl__proplist):
   * subversion/client/status-cmd.c (svn_cl__status):
   * subversion/client/update-cmd.c (svn_cl__update):
   Call push_implicit_dot_target.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@838850 13f79535-47bb-0310-9956-ffa450edef68