*cl.h: added message member to svn_cl__opt_state_t struct
*commit-cmd.c: pass opt_state->message down to
 svn_client_commit. Remove stale comments

*main.c: - alphabetize options[],
         - add --message 'm' option
         - add case for message arg

*tests/svn-test2.sh: added --message to final commit test.

So now messages are getting passed down to the nether regions of the
client libs, but I don't see (from the test) that the message is
finding its way into the xml file generated by the commit. Poked
around a little in libsvn_ra, but decided that sleep was a better
option at this point.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@838924 13f79535-47bb-0310-9956-ffa450edef68