Bug fix for issue #397.
Fix up problems with build/link when --disable-shared is used.

The root problem was the use of SVN_CLIENT_DEPS. This didn't truly handle
the dependencies and link-time behavior properly (in the absence of a
recursive make with ordered subdir builds).

The fix is to recognize that the dependencies are caused by libsvn_ra,
rather than the client. Next, to associate those dependencies directly with
libsvn_ra. libtool provides for this within the .la as long as you depend
directly on .a files rather than .la files (fixed in libtool 1.4). Since we
only need the dependencies during disable-shared, the focus on .a files is
possible. As a result, when programs link against libsvn_ra.la, they
automatically pull in the other libraries (ra_dav and the fs-based libs).

* Makefile.in: toss *_CLIENT_DEPS and define SVN_RA_LIB_DEPS and
    SVN_RA_LIB_LINK. _DEPS for dependency handling and _LINK for what occurs
    within the link line. For static builds, these are filled in. For shared
    builds, they are empty.

    Remove SVN_NEON_LIBS macro (unused).

* gen-make.py: add recognition of add-deps in a definition to add arbitrary
    text to a dependency line.

* build.conf: remove use of SVN_CLIENT_DEPS (see above). expand definition
    of libsvn_ra to conditionally include more dependencies, and to link
    against additional libraries.

* configure.in: no need to SVN_LIBSVN_* so toss the sinclude() of
    svn-define-libs.m4. SVN_NEON_LIBS, SVN_TESTS_MAIN_LIBS, FS_BASED_DIRS,
    and FS_BASED_TESTS are no longer used; remove them. convert definition
    of LIBSVN_CLIENT_DEPS to SVN_RA_LIB_DEPS and SVN_RA_LIB_LINK.

* svn-define-libs.m4: no longer used; removed.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@839434 13f79535-47bb-0310-9956-ffa450edef68