Revamping the organization of the doc/ directory.  Take a look-see in the top-level of README.
As a rule, each texinfo document needs to be isolated in its own
directory (since each doc typically consists of multiple .texi files
and leaves a mess when compiling.)  The Right solution for building
docs is to be able to describe them in our top-level build.conf.  But
for now, we'll just enable our "design" doc to build directly.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@839458 13f79535-47bb-0310-9956-ffa450edef68