Write svndiffs into the strings table, but don't hook them up to representations yet:
* node-rev.c (write_string_baton, write_string): New struct and
svn_write_fn_t func, respectively.
(deltify): Write deltified data to new target string.

* reps-strings.c (svn_fs__rep_read_get_baton): Return the baton.

* reps-strings.h (svn_fs__get_mutable_rep): Doc fix.


In the You've-Got-To-Be-Kidding-Me-Right? Department:

* dag.c (svn_fs__dag_set_contents): Fix confusion between rep keys and
string keys.  You're not going to *believe* the bug that was latent
here:

The fs tests have been passing only because, by sheer coincidence,
they always created exactly the same number of reps and strings, and
did so in such a way that new rep keys were always the same as the new
strings to which the reps referred.  Thus, even though this function
got a rep key ... mistook it for a string key and wrote data to that
string ... then innocently stored the rep key in the node revision
... everything still worked, because the rep and string happened to be
a matching pair anyway!

It all fell apart when deltify() finally started writing svndiff data
into new strings.  Even though the code was not "hooked up" and those
new strings were not yet referred to by any representations, merely
creating them was enough to throw the correspondence off, so files
contents started being randomly wrong.  Whew.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@839668 13f79535-47bb-0310-9956-ffa450edef68