Hot backup script changes.
We're going to run it as a post-commit hook, along with the mail
script.  In addition to backing up the repository, it uses a
write-lock file to "clean" (remove) obsolete logfiles.  (This locking
guards against the case where two commits happen very close together,
whereby the removing of old log files may interfere with another
process' attempt to back them up.)




git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@840121 13f79535-47bb-0310-9956-ffa450edef68