Finish issue #407: "Better subcommand help for command-line client." Specifically, 'svn help subcommand' now prints a list of switches accepted by the subcommand.  (Stay tuned for next commit, which will rewrite the subcommand descriptions themselves.)
Note 1: you'll need to update to the latest APR, or this change won't
compile for you.  (I recently added a field to apr_getopt_option_t.)

Note 2: We now have a fixed-size array of options in each subcommand
structure.  The size of this array is a #define based on the number of
unique options that exist.  Therefore the subcommand structure needs
to be declared right next to the options table, and the options table
cannot live in a header file (lest we get multiple declarations when
linking.)  As a result, everything is declared in main.c now... which
has the effect of effectively moving almost all help-cmd.c code into
main.c (which isn't a big deal, really.)

Note 3:  There is no third thing.

* cl.h (svn_cl__cmd_desc_t):  moved to main.c
  (svn_cl__get_option_from_enum):  declare.
  (svn_cl_-_print_generic_help):  public declaration of formerly
  static func.

* util.c (get_cmd_table_entry):  moved to main.c
  (svn_cl__get_canonical_command):  moved to main.c

* main.c (svn_cl__options): the options table, now made public, and
  add general docstrings to each option structure.

  (print_option):  new static helper func.
  (svn_cl__get_option_from_enum):  new public helper func.

  (svn_cl__cmd_desc_t): moved here from cl.h.  new field: an array of
  accepted option enums.

  (get_cmd_table_entry):  moved here from util.c
  (svn_cl__get_canonical_command):  moved here from util.c

  (svn_cl__cmd_table): for each subcommand structure, fill in new
  field with list of valid switches.  Man, that took a lot of research!
 
  (print_command_info): move here from help-cmd.c, and now print
  options attached to each subcommand.
 
  (svn_cl__print_generic_help): move here from help-cmd.c, renamed
  from static print_generic_help.

  (svn_cl__subcommand_help): move here from help-cmd.c

* help-cmd.c (print_command_info): moved to main.c.
  (svn_cl_subcommand_help): moved to main.c
  (svn_cl__help): update call to svn_cl__print_generic_help.

* status-cmd.c (svn_cl__status): tweak... remove reference to very_verbose.


### VALIDATE!




git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@841013 13f79535-47bb-0310-9956-ffa450edef68