* subversion/libsvn_wc/get_editor.c (make_patch_open_tag): Pass "--silent" to patch program.
Ever since Philip Martin fixed svn_io_run_cmd() in revision 893, the
user sees whatever the patch program prints.  In most cases, patch is
quiet, but when it does hunk adjustment, it chats by default.  See
http://subversion.tigris.org/servlets/ReadMsg?msgId=55760&listName=dev
for more on this situation.

With the --silent flag, patch will only chat when it encounters
problems, such as hunk application failure:

   $ svn up
   G  ./A/D/G/pi
   1 out of 1 hunk FAILED -- saving rejects to rho.61556.00001.rej
   C  ./A/D/G/rho
   U  ./A/D/G/tau
   $ 

Maybe we should leave it like this -- drawing attention to conflicts
isn't such a bad thing.  But it interferes with automated parsing.
Hmmm.




git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@841217 13f79535-47bb-0310-9956-ffa450edef68