This fixes bite-sized task number 598 ("svn status <nonexistent-file>" fails to give an error). Passes stat_test.py just fine.
* subversion/include/svn_wc.h

  (svn_wc_statuses): Added 'svn_boolean_t strict' parameter

* subversion/libsvn_wc/status.c

  (assemble_status): Added 'svn_boolean_t strict' parameter. When
strict is non-zero, return an error if we encounter a path that is not
in the working copy. 
  (add_status_structure): Added 'svn_boolean_t strict'
parameter. Added 'strict' argument to all calls to assemble_status
within.
  (add_unversioned_items): Added FALSE argument to call to assemble_status.
  (svn_wc_status): Added FALSE argument to call to assemble_status.
  (svn_wc_statuses): Added 'svn_boolean_t strict' parameter. Added
'strict' argument to all calls to assemble_status within. Also did
some reformatting to keep within 80 cols.

* subversion/libsvn_client/status.c
  (svn_client_status): Add strict var which is set to TRUE if we're
not updating, and FALSE if we are updating. This allows us to properly
error out if someone tries to stat a file that doesn't exist in the
working copy. Added 'strict' argument to assemble_status invocation.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@841288 13f79535-47bb-0310-9956-ffa450edef68