Fix a bug in Garrett's r1264 commit.
(This was causing commit_test #15 to fail, but this test was written
only to expect generic failure, so it still passing, so it looked as
though 'make check' was passing.  If you ran commit_tests.py by hand,
though, you'd see the extra stderr output.  The moral of the story is
that commit_test #15 should be using run_and_verify_commit() to look for
a *specific* expected errorstring.  Anyone wanna fix it?)

* copy.c (wc_to_wc_copy): make sure it's actually a file you're
  passing to apr_file_remove.  :-)

 

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@841349 13f79535-47bb-0310-9956-ffa450edef68