Convert the update/checkout/status editor to the new svn_delta_editor_t interface.  Let this stand as a warning to any other sucker who casually considers upgrading an editor to the new interface. :-)
* subversion/include/svn_wc.h
  (svn_wc_get_switch_editor, svn_wc_get_checkout_editor,
svn_wc_get_update_editor): Changed 'editor' param to svn_delta_editor_t type.

* subversion/libsvn_wc/update_editor.c
  (make_dir_baton, make_file_baton): Make 'name' a const char *.
  (struct file_baton): Add an 'edit_baton' member.
  (free_file_baton): Removed function.
  (prep_directory): Simplify this function.
  (bump_dir_revision): Was free_dir_baton.  Now doesn't free the dir  baton.
  (make_editor, svn_wc_get_switch_editor, svn_wc_get_checkout_editor,
svn_wc_get_update_editor): Changed 'editor' param to svn_delta_editor_t type.
  (open_root): Add 'pool' argument (and use it).
  (delete_entry, open_directory, open_file): Add 'pool' argument (and use it), 
and make 'name' a const char *.
  (add_directory, add_file): Add 'pool' argument (and use it), and make 'name'
and 'copyfrom_path' const char *'s.
  (add_or_open_file): Add 'pool' argument (and use it), and make
'name' a const char *.  Rename 'ancestor_url' to 'copyfrom_path'
(which is now also a const char *) and 'ancestor_revision' to
'copyfrom_rev'.
  (change_dir_prop): Add 'pool' argument (and use it), make 'name' a
const char *, and 'value' a const svn_string_t *.
  (change_file_prop): Add 'pool' argument, but don't use it (we'll let
our data to live in the same pool as the hash that will contain it).
Make 'name' a const char *, and 'value' a const svn_string_t *.
  (close_edit): No longer free the edit baton's pool.

* subversion/libsvn_client/apply_edits.c
  (apply_delta): Update and checkout editors are now
svn_delta_editor_t, but we will wrap them for compatibility with old
drivers.

* subversion/libsvn_client/checkout.c
  (svn_client_checkout): Checkout editor is now svn_delta_editor_t,
but we will wrap it for compatibility with old drivers.

* subversion/libsvn_client/copy.c
  (repos_to_wc_copy): Checkout editor is now svn_delta_editor_t,
but we will wrap it for compatibility with old drivers.

* subversion/libsvn_client/update.c
  (svn_client_update): Update editor is now svn_delta_editor_t, but we
will wrap it for compatibility with old drivers.

* subversion/libsvn_client/switch.c
  (svn_client_switch): Switch editor is now svn_delta_editor_t, but we
will wrap it for compatibility with old drivers.

* subversion/tests/libsvn_wc/checkout-test.c
  (apply_delta): Checkout editor is now svn_delta_editor_t,
but we will wrap it for compatibility with old drivers.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@841469 13f79535-47bb-0310-9956-ffa450edef68