The FreeBSD port of Subversion goes to some length to link against the system version of expat, rather than the expat-lite we distribute, so it will be a good citizen of the ports tree and avoid spamming over the installed version of expat that already exists.  The patch to build.conf that was accomplishing this was flawed, and only happened to work on my system because I had a copy of libexpat.la sitting around in my /usr/local/lib, I imagine from a previous subversion install I never cleaned out fully.
This alters the patch to build.conf such that instead of just replacing 
libexpat with -lexpat, it replaces it with $(SVN_APRUTIL_LIBS), which in 
addition to -lexpat, will also include -L/usr/local/lib, which makes the 
build work.  It also includes /usr/local/lib/libaprutil.la, which we don't 
strictly need, but it doesn't hurt anything, and will get us by until we 
are smarter about finding expat.

* packages/freebsd/subversion/files/patch-build.conf
  Replace -lexpat with $(SVN_APRUTIL_LIBS), which includes the dir libexpat 
  is installed in, thus making this stuff actually compile on a clean system.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@841645 13f79535-47bb-0310-9956-ffa450edef68