Bugfix:  receiving a copied directory during an update over DAV was failing.  Patch by Ben and CMike.
When mod_dav_svn constructs a vsn-rsc-url, it's not always safe to
append the newest path to the created-rev -- for copied subtrees in
particular.  It's not safe to return "CR/old_path" either.  Thus for
now, we attempt to return "CR/latest_path" as long as the latest path
actually exists in CR and is the same node.  Otherwise, we just return
"target-rev/latest-path".  The only impact of this is that future HTTP
caching will be a teeny bit less efficient.

* util.c (dav_svn_build_uri):  remove useless rev_root arg;  use
  revision arg to build VERSION resource.

  (dav_svn_get_safe_cr): new function to determine if it's safe to
  return "CR/latest_path" instead of "target-rev/latest-path".

* update.c (send_vsn_url): use new dav_svn_get_safe_cr.

* dav_svn.h (dav_svn_build_uri):  no more rev_root arg, update
  docstring requirements for version-urls.

  (dav_svn_get_safe_cr):  new prototype.

* repos.c (dav_svn_prep_version): update call to dav_svn_build_uri.

* version.c (dav_svn_get_option): same.

* liveprops.c (dav_svn_insert_prop): same.  for version-url, use new
  dav_svn_get_safe_cr func.

* merge.c (send_response): same. for version-url, use new
  dav_svn_get_safe_cr func.

  (dav_svn__merge_response): same.

* copy_tests.py (receive_copy_in_update):  new regression test.







git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@842211 13f79535-47bb-0310-9956-ffa450edef68